move FEN notation code to separate file
