+++ src/Util/Util.hs
@@ -1,0 +1,4 @@
+module Util where
+
+-- Returns the first element of a 3-tuple
+getFirst (a, b, c) = a
